requirements.txt
appnope==0.1.0
astroid==2.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.5
django-filter==2.0.0
django-qr-code==1.0.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.12.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.0.0
PyQRCode==1.2.1
python3-openid==3.1.0
pytz==2018.4
qrcode==6.0
requests==2.19.1
requests-oauthlib==1.0.0
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.23
wcwidth==0.1.7
wrapt==1.10.11
